% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckInput.R
\name{check.input}
\alias{check.input}
\title{Check Input for Pooling Data}
\usage{
check.input(frames)
}
\arguments{
\item{frames}{is a list of data frames}
}
\description{
Check Input for Pooling Data
}
\examples{
set.seed(123)
p <- 100
data1 <- data.frame(group=sample(1:3,200,replace=TRUE),matrix(rnorm(p*200),ncol=p))
data2 <- data.frame(group=sample(1:2,150,replace=TRUE),matrix(rnorm(p*150),ncol=p))
frames <- list(data1, data2)
check.input(frames)
}
