% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupDifference.R
\name{GroupDifference}
\alias{GroupDifference}
\title{Pooled test}
\usage{
GroupDifference(x)
}
\arguments{
\item{x}{is a data frame, can be normally distributed or not normally distributed. For 2 groups we will perform a two sample t-test or Wilcoxon rank sum test. For more than 2 groups we will perform ANOVA or Kruskal Wallis test.}
}
\value{
A list containing p values
}
\description{
Pooled test
}
\examples{
set.seed(123)
p <- 100
data1 <- data.frame(group=sample(1:3,200,replace=TRUE),matrix(rnorm(p*200),ncol=p))
GroupDifference(data1)
}
