% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolMin.R
\name{pool.min}
\alias{pool.min}
\title{Min Test Meta Analysis}
\usage{
pool.min(pvalues)
}
\arguments{
\item{input}{is a list of p-values}
}
\value{
the minimum p-value among the K studies as the test statistic
}
\description{
It follows a beta distribution with degrees of freedom alpha=1 and beta=k under the null hypothesis
}
\examples{
pvalues<- c(0.486667826,0.154255057, 0.858419083, 0.486667826 ,0.405013316, 0.316372249, 0.173853520, 0.284391656, 0.370577866, 0.421579844, 0.221750129, 0.962561630, 0.392123661, 0.127102942, 0.896587161, 0.665359191)
pool.min(pvalues)
}
