% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectOverall.R
\name{Project}
\alias{Project}
\title{Pooling Data}
\usage{
Project(x1, x2, x3 = NULL, x4 = NULL, x5 = NULL, test = c("fisher",
  "stouffer", "min", "max"))
}
\arguments{
\item{x1, }{x2, x3, x4, x5 are data frames with each data frame consisting of identical number of columns. Must input at least 2 and at most 5 data frames.}

\item{test}{indicates which pooling test to use, must be either fisher, stouffer, min, max}
}
\value{
P values for the input test
}
\description{
This function uses different tests (Fisher, Stouffer, Min, Max) to pool data from different data frames.
}
\examples{
set.seed(123)
p <- 100
data1 <- data.frame(group=sample(1:3,200,replace=TRUE),matrix(rnorm(p*200),ncol=p))
data2 <- data.frame(group=sample(1:2,150,replace=TRUE),matrix(rnorm(p*150),ncol=p))
Project(x1= data1, x2= data2, test= "fisher")
}
